Address CodeRabbit review on PR #6
- home: delay create-profile modal until the profiles sheet finishes
  dismissing (iOS can't present during a dismissal); guard bonus progress
  against requiredSpend <= 0.
- dates: new localIsoDay() helper replaces toISOString().slice(0,10) for
  today/yesterday (UTC day-shift in US timezones; DST-safe stepping).
- DateField web fallback: re-sync text when value changes externally and
  enforce minimumDate/maximumDate like the native pickers.
- useAddSpendEntry: derive bonus completion from a fresh server-side sum
  instead of client state; a completion-flag failure no longer rejects the
  mutation (retrying after the committed insert duplicated spend rows).
  Full single-transaction RPC deferred to the next schema deploy.
- analytics view: cast redemption_rate division to numeric (future-proofs
  against integer amount columns; migration not yet applied remotely).

+++ supabase/migrations/20260711120000_analytics_views.sql
@@ -98,7 +98,9 @@ select bd.id                              as benefit_definition_id,
        coalesce(r.redeemed_value, 0)      as redeemed_value,
        coalesce(r.redemption_count, 0)    as redemption_count,
        case when coalesce(a.allotted_value, 0) > 0
-            then round(coalesce(r.redeemed_value, 0) / a.allotted_value, 4)
+            -- ::numeric guards against integer division if the amount
+            -- columns ever migrate to integer cents.
+            then round(coalesce(r.redeemed_value, 0)::numeric / a.allotted_value, 4)
             else null end                 as redemption_rate
 from public.benefit_definitions bd
 join public.card_products cp        on cp.id = bd.card_product_id
